chore: cleanup some markdown files
