+++ packages/netlify-cms-widget-markdown/src/MarkdownControl/VisualEditor.js
@@ -47,6 +47,39 @@ const InsertionPoint = styled.div`
   flex: 1 1 auto;
   cursor: text;
 `;
+
+export function mergeMediaConfig(editorComponents, field) {
+  // merge editor media library config to image components
+  if (editorComponents.has('image')) {
+    const imageComponent = editorComponents.get('image');
+    const fields = imageComponent?.fields;
+
+    if (fields) {
+      imageComponent.fields = fields.update(
+        fields.findIndex(f => f.get('widget') === 'image'),
+        f => {
+          // merge `media_library` config
+          if (field.has('media_library')) {
+            f = f.set(
+              'media_library',
+              field.get('media_library').mergeDeep(f.get('media_library')),
+            );
+          }
+          // merge 'media_folder'
+          if (field.has('media_folder') && !f.has('media_folder')) {
+            f = f.set('media_folder', field.get('media_folder'));
+          }
+          // merge 'public_folder'
+          if (field.has('public_folder') && !f.has('public_folder')) {
+            f = f.set('public_folder', field.get('public_folder'));
+          }
+          return f;
+        },
+      );
+    }
+  }
+}
+
 function Editor(props) {
   const {
     onAddAsset,
@@ -73,6 +106,8 @@ function Editor(props) {
     codeBlockComponent || editorComponents.has('code-block')
       ? editorComponents
       : editorComponents.set('code-block', { label: 'Code Block', type: 'code-block' });
+
+  mergeMediaConfig(editorComponents, field);
 
   const [editorValue, setEditorValue] = useState(
     props.value
@@ -141,10 +176,12 @@ function Editor(props) {
 
   const handleDocumentChange = debounce(newValue => {
     setEditorValue(newValue);
-    onChange(slateToMarkdown(newValue, {
-      voidCodeBlock: !!codeBlockComponent,
-      remarkPlugins: getRemarkPlugins(),
-    }));
+    onChange(
+      slateToMarkdown(newValue, {
+        voidCodeBlock: !!codeBlockComponent,
+        remarkPlugins: getRemarkPlugins(),
+      }),
+    );
   }, 150);
 
   function handleChange(newValue) {
@@ -153,7 +190,6 @@ function Editor(props) {
     }
     setToolbarKey(prev => prev + 1);
   }
-
 
   function hasMark(format) {
     return isMarkActive(editor, format);
